+++ packages/node/src/behaviors/operational-credentials/index.ts
@@ -10,3 +10,4 @@ export * from "./OperationalCredentialsInterface.js";
 export * from "./OperationalCredentialsBehavior.js";
 export * from "./OperationalCredentialsServer.js";
 export * from "./OperationalCredentialsClient.js";
+export * from "./VendorIdVerification.js";