+++ packages/protocol/src/certificate/kinds/index.ts
@@ -9,4 +9,5 @@ export * from "./CertificationDeclaration.js";
 export * from "./Icac.js";
 export * from "./Noc.js";
 export * from "./Rcac.js";
+export * from "./Vvsc.js";
 export * from "./X509Base.js";