+++ support/chip-testing/test/core/OPCREDS.test.ts
@@ -5,9 +5,5 @@
  */
 
 describe("OPCREDS", () => {
-    chip("OPCREDS/*").exclude(
-        // These are Matter 1.5
-        "OPCREDS/3.8",
-        "OPCREDS/3.9",
-    );
+    chip("OPCREDS/*");
 });